refactor(simulation): simplify runner protocol
