+++ tests/dt_model/examples/test_mobility_bologna.py
@@ -1,47 +1,48 @@
+# SPDX-License-Identifier: Apache-2.0
 """Regression tests for the Bologna mobility example.
 
 With a fixed random seed the KPI outputs must remain stable across changes.
 """
 
-# SPDX-License-Identifier: Apache-2.0
+from __future__ import annotations
 
 import numpy as np
 import pytest
-from mobility_bologna.mobility_bologna import BolognaModel, compute_kpis, evaluate
+from mobility_bologna.bologna_model import BolognaEvaluator, BolognaModel
 
-# Ensemble size used in all tests. Small enough to be fast, large enough to
-# exercise the ensemble path (distribution-backed indexes are sampled).
+from civic_digital_twins.dt_model import Scenario
+from civic_digital_twins.dt_model.simulation.runner import EvaluationConfig
+
 _ENSEMBLE_SIZE = 5
 
 
 @pytest.fixture(scope="module")
-def model():
+def model() -> BolognaModel:
     """Shared BolognaModel instance (graph construction is the expensive part)."""
     return BolognaModel(**BolognaModel.default_inputs())
 
 
-@pytest.fixture
-def evaluation(model):
-    """Evaluate the model with a fixed seed and return the subs dict."""
-    np.random.seed(42)
-    return evaluate(model, size=_ENSEMBLE_SIZE)
+@pytest.fixture(scope="module")
+def evaluator(model: BolognaModel) -> BolognaEvaluator:
+    """Shared evaluator built on the shared model."""
+    return BolognaEvaluator(model)
 
 
-def test_base_inflow_is_deterministic(model):
+def test_base_inflow_is_deterministic(evaluator: BolognaEvaluator, model: BolognaModel) -> None:
     """Base inflow depends only on fixed data — must be identical across seeds."""
+    config = EvaluationConfig(ensemble_size=_ENSEMBLE_SIZE)
     np.random.seed(0)
-    subs_a = evaluate(model, size=_ENSEMBLE_SIZE)
+    out_a = evaluator.evaluate(Scenario(model), config)
     np.random.seed(99)
-    subs_b = evaluate(model, size=_ENSEMBLE_SIZE)
-    kpis_a = compute_kpis(model, subs_a)
-    kpis_b = compute_kpis(model, subs_b)
-    assert kpis_a["Base inflow [veh/day]"] == kpis_b["Base inflow [veh/day]"]
+    out_b = evaluator.evaluate(Scenario(model), config)
+    assert out_a.kpis["Base inflow [veh/day]"] == out_b.kpis["Base inflow [veh/day]"]
 
 
-def test_kpis_stable_with_fixed_seed(model, evaluation):
+def test_kpis_stable_with_fixed_seed(evaluator: BolognaEvaluator, model: BolognaModel) -> None:
     """All KPIs must reproduce exactly when the random seed is fixed."""
-    kpis = compute_kpis(model, evaluation)
-
+    config = EvaluationConfig(ensemble_size=_ENSEMBLE_SIZE)
+    np.random.seed(42)
+    out = evaluator.evaluate(Scenario(model), config)
     expected = {
         "Base inflow [veh/day]": 168139,
         "Modified inflow [veh/day]": 141044,
@@ -51,5 +52,4 @@ def test_kpis_stable_with_fixed_seed(model, evaluation):
         "Emissions [NOx gr/day]": 508105,
         "Modified emissions [NOx gr/day]": 104427,
     }
-
-    assert kpis == expected
+    assert out.kpis == expected
